fixed bug with converting double to int

+++ src/main/scala/io/prediction/opennlp/engine/Algorithm.scala
@@ -13,8 +13,7 @@ class Algorithm(val ap: AlgorithmParams)
   }
 
   def predict(model: Model, query: Query): PredictedResult = {
-    val sentiment = Sentiment(
-      model.gis.getBestOutcome(model.gis.eval(query.sentence.split(" "))).toInt)
+    val sentiment = Sentiment(model.gis.getBestOutcome(model.gis.eval(query.sentence.split(" "))).toDouble.toInt)
     PredictedResult(sentiment.toString)
   }
